+++ QuickFIXn/FixValues.cs
@@ -4,21 +4,20 @@ namespace QuickFix
 {
     public class FixValue<T>
     {
-        private T value_;
-        private string description_;
+        private T _value;
 
-        public T Value { get { return value_; } }
-        public string Description { get { return description_; } }
+        public T Value => _value;
+        public string Description { get; }
 
         public FixValue(T value, string description)
         {
-            value_ = value;
-            description_ = description;
+            _value = value;
+            Description = description;
         }
 
-        public override bool Equals(object obj)
+        public override bool Equals(object? obj)
         {
-            if ((null == obj) || (this.GetType() != obj.GetType()))
+            if (obj is null || (this.GetType() != obj.GetType()))
                 return false;
             FixValue<T> rhs = (FixValue<T>)obj;
             return this.Value.Equals(rhs.Value);
@@ -26,12 +25,12 @@ namespace QuickFix
 
         public override int GetHashCode()
         {
-            return value_.GetHashCode();
+            return _value.GetHashCode();
         }
 
         public override string ToString()
         {
-            return description_;
+            return Description;
         }
     }
 
@@ -39,8 +38,6 @@ namespace QuickFix
     {
         public static class ApplVerID
         {
-            public const string FIX27 = "0";
-            public const string FIX30 = "1";
             public const string FIX40 = "2";
             public const string FIX41 = "3";
             public const string FIX42 = "4";
@@ -52,46 +49,34 @@ namespace QuickFix
 
             public static string FromBeginString(string beginString)
             {
-                if (BeginString.FIX40.Equals(beginString))
-                    return ApplVerID.FIX40;
-                else if (BeginString.FIX41.Equals(beginString))
-                    return ApplVerID.FIX41;
-                else if (BeginString.FIX42.Equals(beginString))
-                    return ApplVerID.FIX42;
-                else if (BeginString.FIX43.Equals(beginString))
-                    return ApplVerID.FIX43;
-                else if (BeginString.FIX44.Equals(beginString))
-                    return ApplVerID.FIX44;
-                else if (BeginString.FIX50.Equals(beginString))
-                    return ApplVerID.FIX50;
-                else if (BeginString.FIX50SP1.Equals(beginString))
-                    return ApplVerID.FIX50SP1;
-                else if (BeginString.FIX50SP2.Equals(beginString))
-                    return ApplVerID.FIX50SP2;
-                else
-                    return beginString;
+                return beginString switch
+                {
+                    BeginString.FIX40 => FIX40,
+                    BeginString.FIX41 => FIX41,
+                    BeginString.FIX42 => FIX42,
+                    BeginString.FIX43 => FIX43,
+                    BeginString.FIX44 => FIX44,
+                    BeginString.FIX50 => FIX50,
+                    BeginString.FIX50SP1 => FIX50SP1,
+                    BeginString.FIX50SP2 => FIX50SP2,
+                    _ => beginString
+                };
             }
 
-            public static string ToBeginString(string applVerId)
-            {
-                if (applVerId == ApplVerID.FIX40)
-                    return BeginString.FIX40;
-                else if (applVerId == ApplVerID.FIX41)
-                    return BeginString.FIX41;
-                else if (applVerId == ApplVerID.FIX42)
-                    return BeginString.FIX42;
-                else if (applVerId == ApplVerID.FIX43)
-                    return BeginString.FIX43;
-                else if (applVerId == ApplVerID.FIX44)
-                    return BeginString.FIX44;
-                else if (applVerId == ApplVerID.FIX50)
-                    return BeginString.FIX50;
-                else if (applVerId == ApplVerID.FIX50SP1)
-                    return BeginString.FIX50SP1;
-                else if (applVerId == ApplVerID.FIX50SP2)
-                    return BeginString.FIX50SP2;
-                else
-                    throw new System.ArgumentException($"ApplVerId parameter '{applVerId}' does not map to a known BeginString");
+            public static string ToBeginString(string applVerId) {
+                return applVerId switch
+                {
+                    FIX40 => BeginString.FIX40,
+                    FIX41 => BeginString.FIX41,
+                    FIX42 => BeginString.FIX42,
+                    FIX43 => BeginString.FIX43,
+                    FIX44 => BeginString.FIX44,
+                    FIX50 => BeginString.FIX50,
+                    FIX50SP1 => BeginString.FIX50SP1,
+                    FIX50SP2 => BeginString.FIX50SP2,
+                    _ => throw new System.ArgumentException(
+                        $"ApplVerId parameter '{applVerId}' does not map to a known BeginString")
+                };
             }
         }
 
@@ -112,25 +97,25 @@ namespace QuickFix
 
         public class SessionRejectReason : FixValue<int>
         {
-            public static SessionRejectReason INVALID_TAG_NUMBER = new SessionRejectReason(0, "Invalid tag number");
-            public static SessionRejectReason REQUIRED_TAG_MISSING = new SessionRejectReason(1, "Required tag missing");
-            public static SessionRejectReason TAG_NOT_DEFINED_FOR_THIS_MESSAGE_TYPE = new SessionRejectReason(2, "Tag not defined for this message type");
-            public static SessionRejectReason UNDEFINED_TAG = new SessionRejectReason(3, "Undefined Tag");
-            public static SessionRejectReason TAG_SPECIFIED_WITHOUT_A_VALUE = new SessionRejectReason(4, "Tag specified without a value");
-            public static SessionRejectReason VALUE_IS_INCORRECT = new SessionRejectReason(5, "Value is incorrect (out of range) for this tag");
-            public static SessionRejectReason INCORRECT_DATA_FORMAT_FOR_VALUE = new SessionRejectReason(6, "Incorrect data format for value");
-            public static SessionRejectReason DECRYPTION_PROBLEM = new SessionRejectReason(7, "Decryption problem");
-            public static SessionRejectReason SIGNATURE_PROBLEM = new SessionRejectReason(8, "Signature problem");
-            public static SessionRejectReason COMPID_PROBLEM = new SessionRejectReason(9, "CompID problem");
-            public static SessionRejectReason SENDING_TIME_ACCURACY_PROBLEM = new SessionRejectReason(10, "SendingTime accuracy problem");
-            public static SessionRejectReason INVALID_MSGTYPE = new SessionRejectReason(11, "Invalid MsgType");
-            public static SessionRejectReason XML_VALIDATION_ERROR = new SessionRejectReason(12, "XML validation error");
-            public static SessionRejectReason TAG_APPEARS_MORE_THAN_ONCE = new SessionRejectReason(13, "Tag appears more than once");
-            public static SessionRejectReason TAG_SPECIFIED_OUT_OF_REQUIRED_ORDER = new SessionRejectReason(14, "Tag specified out of required order");
-            public static SessionRejectReason REPEATING_GROUP_FIELDS_OUT_OF_ORDER = new SessionRejectReason(15, "Repeating group fields out of order");
-            public static SessionRejectReason INCORRECT_NUM_IN_GROUP_COUNT_FOR_REPEATING_GROUP = new SessionRejectReason(16, "Incorrect NumInGroup count for repeating group");
-            public static SessionRejectReason NON_DATA_VALUE_INCLUDES_FIELD_DELIMITER = new SessionRejectReason(17, "Non-data value includes field delimiter");
-            public static SessionRejectReason OTHER = new SessionRejectReason(99, "Other");
+            public static SessionRejectReason INVALID_TAG_NUMBER = new(0, "Invalid tag number");
+            public static SessionRejectReason REQUIRED_TAG_MISSING = new(1, "Required tag missing");
+            public static SessionRejectReason TAG_NOT_DEFINED_FOR_THIS_MESSAGE_TYPE = new(2, "Tag not defined for this message type");
+            public static SessionRejectReason UNDEFINED_TAG = new(3, "Undefined Tag");
+            public static SessionRejectReason TAG_SPECIFIED_WITHOUT_A_VALUE = new(4, "Tag specified without a value");
+            public static SessionRejectReason VALUE_IS_INCORRECT = new(5, "Value is incorrect (out of range) for this tag");
+            public static SessionRejectReason INCORRECT_DATA_FORMAT_FOR_VALUE = new(6, "Incorrect data format for value");
+            public static SessionRejectReason DECRYPTION_PROBLEM = new(7, "Decryption problem");
+            public static SessionRejectReason SIGNATURE_PROBLEM = new(8, "Signature problem");
+            public static SessionRejectReason COMPID_PROBLEM = new(9, "CompID problem");
+            public static SessionRejectReason SENDING_TIME_ACCURACY_PROBLEM = new(10, "SendingTime accuracy problem");
+            public static SessionRejectReason INVALID_MSGTYPE = new(11, "Invalid MsgType");
+            public static SessionRejectReason XML_VALIDATION_ERROR = new(12, "XML validation error");
+            public static SessionRejectReason TAG_APPEARS_MORE_THAN_ONCE = new(13, "Tag appears more than once");
+            public static SessionRejectReason TAG_SPECIFIED_OUT_OF_REQUIRED_ORDER = new(14, "Tag specified out of required order");
+            public static SessionRejectReason REPEATING_GROUP_FIELDS_OUT_OF_ORDER = new(15, "Repeating group fields out of order");
+            public static SessionRejectReason INCORRECT_NUM_IN_GROUP_COUNT_FOR_REPEATING_GROUP = new(16, "Incorrect NumInGroup count for repeating group");
+            public static SessionRejectReason NON_DATA_VALUE_INCLUDES_FIELD_DELIMITER = new(17, "Non-data value includes field delimiter");
+            public static SessionRejectReason OTHER = new(99, "Other");
 
             public SessionRejectReason(int value, string description)
                 : base(value, description)